feat: refresh token fixed. Loop on checking token
