+++ core/data/src/main/java/com/juandgaines/core/data/di/SessionManagerModule.kt
@@ -3,9 +3,7 @@ package com.juandgaines.core.data.di
 import android.content.SharedPreferences
 import com.juandgaines.core.data.auth.SharedPreferencesSessionManager
 import com.juandgaines.core.data.auth.TokenApi
-import com.juandgaines.core.data.auth.refresh_token.RemoteTokenDataSourceImpl
 import com.juandgaines.core.domain.auth.SessionManager
-import com.juandgaines.core.domain.auth.refresh_token.RemoteTokenDataSource
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -33,14 +31,6 @@ class SessionManagerModule {
         retrofit: Retrofit
     ): TokenApi {
         return retrofit.create(TokenApi::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideRemoteTokenDataSource(
-        tokenApi: TokenApi
-    ): RemoteTokenDataSource {
-        return RemoteTokenDataSourceImpl(tokenApi)
     }
 }
